removed rules for builtin/env.c and .h from Makefile and added inclusion of these to builtin/builtin.c and h
